extracted the notifcaiton  component
